Moved functionality from upload.py to assignments.py and delted upload.py
